# runtime/internal/agentapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sonalmod agent runtime API
  description: |
    HTTP API for agent runs backed by the ADK `AgentRunner` / `Run` flow in this module.
    The runner uses `StreamingModeSSE`; HTTP responses use **Server-Sent Events** (`text/event-stream`).

    **Stream contract:** Each SSE message uses standard framing (`event:` optional, `data:` JSON).
    Every `data` line is a JSON object with an `event` discriminator (see `StreamEvent`).

    **Start vs continue:** `POST /agent-runs` begins a new session; the server assigns `sessionId`
    and emits it in the first `sessionBound` event. `POST /sessions/{sessionId}/agent-runs`
    continues an existing session (same request body shape; session comes from the path).
  version: 0.1.0
servers:
  - url: /
    description: Relative to the HTTP server base URL
...
paths:
...
  /agent-runs:
    post:
      tags: [Agent runs]
      operationId: startAgentRun
      summary: Start a new agent run
      description: |
        Creates a new session, runs the agent with the given user message, and streams events over SSE.
        The assigned `sessionId` appears in the first `sessionBound` event in the stream.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: SSE stream of agent events until the run completes or errors.
          headers:
            Content-Type:
              schema:
                type: string
                default: text/event-stream; charset=utf-8
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  UTF-8 SSE stream. Each event’s `data` field is JSON matching `StreamEvent`
                  (see examples and components).
              examples:
                sample:
                  value: |
                    event: sessionBound
                    data: {"event":"sessionBound","sessionId":"550e8400-e29b-41d4-a716-446655440000"}

                    event: agent
                    data: {"event":"agent","partial":true,"content":{"parts":[{"text":"Hello"}]}}

                    event: agent
                    data: {"event":"agent","partial":false,"turnComplete":true,"content":{"parts":[{"text":"Hello, world."}]}}

                    event: done
                    data: {"event":"done"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /sessions/{sessionId}:
    get:
      tags: [Agent runs]
      operationId: readSession
      summary: Read session events (replay history and optionally tail live run)
      description: |
        Replays all historical session events as SSE frames. If a run is currently active for the
        session, the stream stays open and continues forwarding live events until the run completes.
        If the session is idle, the stream ends with a `done` event after replaying history.

        Stream order:
        1. `sessionBound` event (confirms the session id)
        2. `sessionStatus` event (`active` or `idle`)
        3. Zero or more `agent` events replayed from history
        4. If active: continues with live `agent` events until the run finishes
        5. `done` event
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: SSE stream replaying session events; stays open if a run is active.
          headers:
            Content-Type:
              schema:
                type: string
                default: text/event-stream; charset=utf-8
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  UTF-8 SSE stream. Each event's `data` field is JSON matching `StreamEvent`.
        '404':
          description: Session not found.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /sessions/{sessionId}/agent-runs:
    post:
      tags: [Agent runs]
      operationId: continueAgentRun
      summary: Continue an agent run for an existing session
      description: |
        Sends a new user message in the given session and streams the agent response over SSE.
        Same request body as `startAgentRun`; `sessionId` is taken from the path (from a prior
        `sessionBound` event or a client-chosen id, depending on server policy).
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: SSE stream of agent events until the run completes or errors.
          headers:
            Content-Type:
              schema:
                type: string
                default: text/event-stream; charset=utf-8
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                sample:
                  value: |
                    event: agent
                    data: {"event":"agent","partial":true,"content":{"parts":[{"text":"Sure —"}]}}

                    event: done
                    data: {"event":"done"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: |
            Returned only if the server validates that the session exists before running.
            Implementations that always resolve sessions (get-or-create) may omit this response.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          $ref: '#/components/responses/InternalError'
  /models:
    get:
      tags: [Models]
      operationId: listModels
      summary: List all available models across all providers
      description: |
        Returns all models configured across all providers. Each entry includes the provider name
        to allow constructing the fully-qualified model name (`provider/model-name`) for agent runs.
      responses:
        '200':
          description: List of available models.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: Session identifier (e.g. UUID) returned from a previous `sessionBound` or chosen for a new conversation.
      schema:
        type: string
        minLength: 1
...
  schemas:
    AgentRunRequest:
      type: object
      required: [message]
      properties:
        message:
          $ref: '#/components/schemas/UserMessageContent'
        model:
          type: string
          description: Optional model name; empty selects the runtime default (`NewAgentRunnerParams.ModelName`).
      additionalProperties: false
    UserMessageContent:
      description: |
        User message for `POST /agent-runs` / continue run: text parts (always user role; not sent by clients).
      type: object
      properties:
        parts:
          type: array
          items:
            $ref: '#/components/schemas/UserMessagePart'
          description: |
            Must contain at least one segment with non-empty text after server-side validation; empty lists are rejected for POST bodies.
      required: [parts]
      additionalProperties: false
    UserMessagePart:
      type: object
      description: Single text segment in a user message (request body).
      properties:
        text:
          type: string
      required: [text]
      additionalProperties: false
...
    ModelInfo:
      type: object
      description: A model available across providers, including the provider name for fully-qualified identification.
      required: [provider, name]
      properties:
        provider:
          type: string
          description: Provider name (prefix in fully-qualified model names like `provider/model-name`).
        name:
          type: string
          description: Technical model identifier (e.g., "gpt-4.1").
        displayName:
          type: string
          description: Optional human-friendly label.
      additionalProperties: false
    ModelListResponse:
      type: object
      description: List of available models across all providers.
      required: [models]
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelInfo'
      additionalProperties: false
...
    ProblemDetails:
      type: object
      description: RFC 9457 problem details (subset).
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
      additionalProperties: true
  responses:
    BadRequest:
      description: Malformed request body or validation failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Authentication required or invalid credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    InternalError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'